% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{draw_graph}
\alias{draw_graph}
\title{Draw the network graph}
\usage{
draw_graph(
  graph,
  layout,
  edges,
  name,
  gsea_result,
  low = low,
  mid = mid,
  high = high
)
}
\arguments{
\item{graph}{Graph object created by ggraph}

\item{layout}{Layouts types available in the ggraph package or created with create_layout()}

\item{edges}{Table with information about the edges. Output of make_edge().}

\item{name}{Can be either "jaccard similarity" or "intersect"}

\item{gsea_result}{A Data Frame of a gene-set enrichment analysis result. Required columns: 
Description, NES, core_enrichment.}

\item{low}{Color of the gene-sets with negative NES. Default color:"#1465AC"}

\item{mid}{Color of the gene-sets with zero NES. Default color: "white"}

\item{high}{Color of the gene-sets with positive NES. Default color: "#B31B21"}
}
\value{
Returns a ggplot object onto which additional layers can be added.
}
\description{
Creates a network of the gene-sets.
}
