% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{make_edge}
\alias{make_edge}
\title{Make the edges of the graph}
\usage{
make_edge(df, name, cutoff, gsea_result)
}
\arguments{
\item{df}{Output of make_jaccard_similarity_table() or make_intersection_table()}

\item{name}{"jaccard similarity" or "intersect"}

\item{cutoff}{Cutoff value of intersections/jaccard similarity index}

\item{gsea_result}{A Data Frame of a gene-set enrichment analysis result. Required columns: 
Description, NES, core_enrichment.}
}
\value{
Returns a table that defines the edges of the graph. It holds the 
names of the two nodes an edge will connect and a value 
- intersection or jaccard similarity - that will determine the line_width
}
\description{
Creates a table that holds information about the edges of the graph. We can 
specify "intersect" or "jaccard similarity" as the value that will determine 
the width of the connecting lines on the graph.
}
