% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{make_intersection_table}
\alias{make_intersection_table}
\title{Calculates the number of intersections}
\usage{
make_intersection_table(gsea_result)
}
\arguments{
\item{set_gene}{A Data Frame object, output of the make_set_gene() function}
}
\description{
Takes a gene-set enrichment analysis result as input and calculates the 
number of intersections between each pair of gene-sets
}
