% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{make_jaccard_similarity_table}
\alias{make_jaccard_similarity_table}
\title{Calculates Jaccard similarity between each pair of gene-sets}
\usage{
make_jaccard_similarity_table(gsea_result)
}
\arguments{
\item{set_gene}{A Data Frame object, output of the make_set_gene() function}
}
\value{
Returns a Data Frame object. Each column and row represents a gene-set
and values are the Jaccard similarity indexes of two gene-sets.
}
\description{
Takes a gene-set enrichment analysis result as input and calculates the 
Jaccard similarity of every gene-set pair.
}
