% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{make_set_gene}
\alias{make_set_gene}
\title{Shows whether a gene is in a gene-set}
\usage{
make_set_gene(gsea_result)
}
\arguments{
\item{gsea_result}{A Data Frame of a gene-set enrichment analysis result. Required columns: 
Description, NES, core_enrichment.}
}
\description{
Takes a gene-set enrichment analysis result as input and creates a 
Data Frame that holds every gene-set provided in the input (columns) and 
every gene (rows) that is a part of these gene-sets. When a gene is part of a 
gene-set the corresponding cell contains 1, when the gene isn't part of the 
gene-set it contains 0.
}
